refactor: changing the order of folders
