--- src/main/java/com/example/cloudservice/controller/autentication/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.example.cloudservice.controller.autentication;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ErrorResponse implements Response{
-    private String message;
-    private final int id = 0;
-}
