--- src/main/java/com/example/cloudservice/controller/autentication/RegisterRequest.java
@@ -1,18 +1,0 @@
-package com.example.cloudservice.controller.autentication;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-
-    private String firstname;
-    private String lastname;
-    private String email;
-    private String password;
-}
