--- src/main/java/com/example/cloudservice/controller/autentication/Response.java
@@ -1,4 +1,0 @@
-package com.example.cloudservice.controller.autentication;
-
-public interface Response {
-}
