+++ src/main/java/com/example/cloudservice/controller/request/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.example.cloudservice.controller.autentication;
+package com.example.cloudservice.controller.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationResponse implements Response{
-    @JsonProperty("auth-token")
-    private String token;
+public class  AuthenticationRequest {
+    @JsonProperty("login")
+    private String email;
+    private String password;
 }
