+++ src/main/java/com/example/cloudservice/controller/response/Response.java
@@ -1,0 +1,4 @@
+package com.example.cloudservice.controller.response;
+
+public interface Response {
+}
